refrectaor de la classe dans le model. Passage de quelque variable en camelCase
